chore: stage section improvement
